refactor(go/ai): consolidate format code into format.go
Merge formatter.go and the per-format files (text, json, jsonl, array, enum)
into a single format.go, and rename formatter_test.go to format_test.go,
matching the V2 branch layout. No behavior change.
