+++ go/ai/format_test.go
